Conflicts:
	src/model/gizmo/Gizmo.java

+++ src/model/Model.java
@@ -51,7 +51,6 @@ public class Model extends Observable implements Board {
 		connectList = new ArrayList<ArrayList<Object>>();
 
 		rotateMap = new HashMap<String, Integer>();
-
 	}
 
 	public void tick() {
@@ -253,7 +252,7 @@ public class Model extends Observable implements Board {
 		notifyObservers();
 	}
 
-	public String addBall(String id, int x, int y) {
+	public String addBall(String id, double x, double y) {
 		if (id == null) {
 			id = generateId("B");
 		}
